+++ datatig/sqlite.py
@@ -74,6 +74,7 @@ class DataStoreSQLite:
                 title TEXT,
                 description TEXT,
                 sort INTEGER NOT NULL,
+                required INTEGER NOT NULL,
                 extra_config TEXT,
                 PRIMARY KEY(type_id, id),
                 FOREIGN KEY(type_id) REFERENCES type(id)
@@ -149,8 +150,8 @@ class DataStoreSQLite:
     def _create_type_add_field(self, cur, type, type_field, parent_keys=[]):
         cur.execute(
             """INSERT INTO type_field (
-            type_id , id, key, type, title, description, sort, extra_config
-            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?)""",
+            type_id , id, key, type, title, description, sort, required, extra_config
+            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)""",
             [
                 type.get_id(),
                 "/".join(parent_keys + [type_field.get_id()]),
@@ -159,6 +160,7 @@ class DataStoreSQLite:
                 type_field.get_title(),
                 type_field.get_description(),
                 self._type_field_sort,
+                1 if type_field.get_required() else 0,
                 json.dumps(type_field.get_extra_config()),
             ],
         )