+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Datatig currently assumes `datatig.yml` is at the root of a git repository; allow to be in subdirs with new `githost/directory` option https://github.com/DataTig/DataTig/issues/53
 * `versionedcheck` CLI now shows errors where files could not be parsed (previously only showed validation errors where files could be parsed)
+* `required` option to all field types https://github.com/DataTig/DataTig/issues/42
+* ... and `required` column to `type_field` table in SQLite database
 
 Internal Python API: